Re-merge integration_spec
